assn3/requirements.txt
tensorflow==1.13.1
tensorflow-estimator==1.13.0
numpy==1.16.2
Pillow==5.4.1
keras==2.2.4
